+++ test/unit/mTokenGateway/mTokenGateway_mint.t.sol
@@ -6,8 +6,6 @@ import {ImTokenOperationTypes} from "src/interfaces/ImToken.sol";
 
 import {mToken_Unit_Shared} from "../shared/mToken_Unit_Shared.t.sol";
 
-import {mTokenGateway} from "src/mToken/extension/mTokenGateway.sol";
-
 contract mTokenGateway_mint is mToken_Unit_Shared {
     function test_RevertWhen_AmountIs0() external {
         // it should revert
@@ -55,11 +53,15 @@ contract mTokenGateway_mint is mToken_Unit_Shared {
 
         // it should increase nonce for this operation type
         assertEq(
-            mWethExtension.getNonce(address(this), block.chainid, ImTokenOperationTypes.OperationType.MintOnOtherChain),
+            mWethExtension.getNonce(
+                address(this), uint32(block.chainid), ImTokenOperationTypes.OperationType.MintOnOtherChain
+            ),
             1
         );
 
         // it should not increase nonce for any other operation type
-        assertEq(mWethExtension.getNonce(address(this), block.chainid, ImTokenOperationTypes.OperationType.Borrow), 0);
+        assertEq(
+            mWethExtension.getNonce(address(this), uint32(block.chainid), ImTokenOperationTypes.OperationType.Borrow), 0
+        );
     }
 }
