+++ test/unit/mTokenGateway/mTokenGateway_withdraw.t.sol
@@ -55,13 +55,15 @@ contract mTokenGateway_withdraw is mToken_Unit_Shared {
         // it should increase nonce for this operation type
         assertEq(
             mWethExtension.getNonce(
-                address(this), block.chainid, ImTokenOperationTypes.OperationType.RedeemOnOtherChain
+                address(this), uint32(block.chainid), ImTokenOperationTypes.OperationType.RedeemOnOtherChain
             ),
             1
         );
 
         // it should not increase nonce for any other operation type
-        assertEq(mWethExtension.getNonce(address(this), block.chainid, ImTokenOperationTypes.OperationType.Borrow), 0);
+        assertEq(
+            mWethExtension.getNonce(address(this), uint32(block.chainid), ImTokenOperationTypes.OperationType.Borrow), 0
+        );
 
         // it should decrease total supply
         assertEq(totalSupplyAfter + amount, totalSupplyBefore);
